% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{experiment}
\alias{experiment}
\alias{exper}
\title{Experiment data set}
\usage{
data(experiment)
}
\value{
A data frame
}
\description{
This data frame dummy_data represents an experiment, 
where we have ten treatments:
A, B, ..., J and measurements at some time points. 
We want to create a scatter plot of measurement against time, 
for each treatment type.
}
\examples{
data(experiment)
}
\keyword{datasets}
