% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrRssprogramming-package.R
\docType{package}
\name{jrRssprogramming-package}
\alias{jrRssprogramming-package}
\alias{jrRssprogramming}
\title{R courses from Jumping Rivers}
\description{
Functions and datasets used in courses by Jumping Rivers.
}
\keyword{package}
